Move runtime_configuration to core/
