+++ core/specfem/program/program.cpp
@@ -4,10 +4,10 @@
 #include "io/interface.hpp"
 #include "kokkos_abstractions.h"
 #include "mesh/mesh.hpp"
-#include "parameter_parser/interface.hpp"
 #include "specfem/assembly.hpp"
 #include "specfem/logger.hpp"
 #include "specfem/receivers.hpp"
+#include "specfem/runtime_configuration.hpp"
 #include "specfem/solver.hpp"
 #include "specfem/source.hpp"
 #include "specfem/timescheme.hpp"
