+++ core/specfem/runtime_configuration.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "runtime_configuration/database_configuration.hpp"
+#include "runtime_configuration/header.hpp"
+#include "runtime_configuration/quadrature.hpp"
+#include "runtime_configuration/run_setup.hpp"
+#include "runtime_configuration/setup.hpp"
+// #include "specfem/solver/interface.hpp"
+#include "io/seismogram/writer.hpp"
+#include "io/wavefield/writer.hpp"
+#include "runtime_configuration/time_scheme.hpp"
