+++ core/specfem/runtime_configuration/database_configuration.hpp
@@ -1,5 +1,4 @@
-#ifndef _PARAMETER_RUNTIME_CONFIGURATION_HPP
-#define _PARAMETER_RUNTIME_CONFIGURATION_HPP
+#pragma once
 
 #include "specfem_setup.hpp"
 #include "yaml-cpp/yaml.h"
@@ -38,5 +37,3 @@ private:
 
 } // namespace runtime_configuration
 } // namespace specfem
-
-#endif
