+++ core/specfem/runtime_configuration/elastic_wave.hpp
