+++ core/specfem/runtime_configuration/electromagnetic_wave.hpp
