+++ core/specfem/runtime_configuration/header.cpp
@@ -1,4 +1,4 @@
-#include "parameter_parser/header.hpp"
+#include "header.hpp"
 #include "yaml-cpp/yaml.h"
 #include <ostream>
