+++ core/specfem/runtime_configuration/header.hpp
@@ -1,5 +1,4 @@
-#ifndef _PARAMETER_HEADER_HPP
-#define _PARAMETER_HEADER_HPP
+#pragma once
 
 #include "specfem_setup.hpp"
 #include "yaml-cpp/yaml.h"
@@ -52,5 +51,3 @@ private:
 };
 } // namespace runtime_configuration
 } // namespace specfem
-
-#endif
