+++ core/specfem/runtime_configuration/quadrature.cpp
@@ -1,4 +1,4 @@
-#include "parameter_parser/quadrature.hpp"
+#include "quadrature.hpp"
 #include "specfem/quadrature.hpp"
 #include "specfem/utilities.hpp"
 #include "yaml-cpp/yaml.h"
