+++ core/specfem/runtime_configuration/run_setup.cpp
@@ -1,4 +1,4 @@
-#include "parameter_parser/run_setup.hpp"
+#include "run_setup.hpp"
 #include "yaml-cpp/yaml.h"
 #include <ostream>
