+++ core/specfem/runtime_configuration/solver.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "solver/solver.hpp"
+#include "solver/solver.tpp"
