+++ core/specfem/runtime_configuration/sources.hpp
