+++ core/specfem/runtime_configuration/time_scheme.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+// #include "specfem/solver.hpp"
+#include "time_scheme/time_scheme.hpp"
