+++ core/specfem/runtime_configuration/writer/kernel.hpp
