+++ core/specfem/runtime_configuration/writer/property.hpp
