+++ core/specfem/runtime_configuration/writer/seismogram.cpp
@@ -1,4 +1,4 @@
-#include "parameter_parser/writer/seismogram.hpp"
+#include "seismogram.hpp"
 #include "io/seismogram/writer.hpp"
 #include "specfem/utilities.hpp"
 #include "yaml-cpp/yaml.h"
