+++ core/specfem/runtime_configuration/writer/wavefield.cpp
@@ -1,4 +1,4 @@
-#include "parameter_parser/writer/wavefield.hpp"
+#include "wavefield.hpp"
 #include "io/ADIOS2/ADIOS2.hpp"
 #include "io/ASCII/ASCII.hpp"
 #include "io/HDF5/HDF5.hpp"
