--- include/parameter_parser/interface.hpp
@@ -1,14 +1,0 @@
-#ifndef _PARAMETER_PARSER_HPP
-#define _PARAMETER_PARSER_HPP
-
-#include "database_configuration.hpp"
-#include "header.hpp"
-#include "quadrature.hpp"
-#include "run_setup.hpp"
-#include "setup.hpp"
-// #include "specfem/solver/interface.hpp"
-#include "io/seismogram/writer.hpp"
-#include "io/wavefield/writer.hpp"
-#include "time_scheme/interface.hpp"
-
-#endif
