--- include/parameter_parser/solver/interface.hpp
@@ -1,7 +1,0 @@
-#ifndef _SPECFEM_RUNTIME_CONFIGURATION_SOLVER_INTERFACE_HPP_
-#define _SPECFEM_RUNTIME_CONFIGURATION_SOLVER_INTERFACE_HPP_
-
-#include "solver.hpp"
-#include "solver.tpp"
-
-#endif // _SPECFEM_RUNTIME_CONFIGURATION_SOLVER_INTERFACE_HPP_
