--- include/parameter_parser/time_scheme/interface.hpp
@@ -1,7 +1,0 @@
-#ifndef _PARAMETER_SOLVER_INTERFACE_HPP
-#define _PARAMETER_SOLVER_INTERFACE_HPP
-
-// #include "specfem/solver.hpp"
-#include "time_scheme.hpp"
-
-#endif
