+++ tests/unit-tests/CMakeLists.txt
@@ -794,7 +794,7 @@ target_link_libraries(
   yaml-cpp
   specfem_environment
   assembly
-  parameter_reader
+  runtime_configuration
   compare_arrays
   timescheme
   point
@@ -820,7 +820,7 @@ target_link_libraries(
   yaml-cpp
   specfem_environment
   assembly
-  parameter_reader
+  runtime_configuration
   compare_arrays
   timescheme
   point
@@ -844,7 +844,7 @@ target_link_libraries(
 #   yaml-cpp
 #   specfem_environment
 #   assembly
-#   parameter_reader
+#   runtime_configuration
 #   writer
 #   periodic_tasks
 #   domain
@@ -865,7 +865,7 @@ target_link_libraries(
 #   yaml-cpp
 #   specfem_environment
 #   assembly
-#   parameter_reader
+#   runtime_configuration
 #   writer
 #   periodic_tasks
 #   domain
