+++ tests/unit-tests/displacement_tests/Newmark/dim2/newmark_tests.cpp
@@ -4,10 +4,10 @@
 #include "io/interface.hpp"
 #include "io/seismogram/reader.hpp"
 #include "mesh/mesh.hpp"
-#include "parameter_parser/interface.hpp"
 #include "specfem/assembly.hpp"
 #include "specfem/logger.hpp"
 #include "specfem/quadrature.hpp"
+#include "specfem/runtime_configuration.hpp"
 #include "specfem/solver.hpp"
 #include "specfem/timescheme.hpp"
 #include "yaml-cpp/yaml.h"
