+++ tests/unit-tests/displacement_tests/Newmark/dim3/newmark_tests.cpp
@@ -5,10 +5,10 @@
 #include "io/seismogram/reader.hpp"
 #include "io/seismogram/writer.hpp"
 #include "mesh/mesh.hpp"
-#include "parameter_parser/interface.hpp"
 #include "specfem/assembly.hpp"
 #include "specfem/logger.hpp"
 #include "specfem/quadrature.hpp"
+#include "specfem/runtime_configuration.hpp"
 #include "specfem/solver.hpp"
 #include "specfem/timescheme.hpp"
 #include "specfem/utilities.hpp"
