+++ tests/unit-tests/seismogram/acoustic/seismogram_tests.cpp
@@ -5,11 +5,11 @@
 #include "io/fortranio/interface.hpp"
 #include "io/interface.hpp"
 #include "mesh/mesh.hpp"
-#include "parameter_parser/interface.hpp"
 #include "specfem/assembly.hpp"
 #include "specfem/logger.hpp"
 #include "specfem/mpi.hpp"
 #include "specfem/quadrature.hpp"
+#include "specfem/runtime_configuration.hpp"
 #include "specfem/solver.hpp"
 #include "specfem/timescheme.hpp"
 #include "yaml-cpp/yaml.h"
